Port KCM to new metadata approach
De-duplicate json metadata for package and C++ plugin (https://phabricator.kde.org/T14983).
Convert the metadata for the KCM to JSON (https://phabricator.kde.org/T14564).
Also install the KCm in the namespace for systemsettings, for the integration with the task manager
a desktop file with NoDisplay=true is shipped (https://phabricator.kde.org/T14517).