kcm: Fix usage of Kirigami.Theme colors
Kirigami.Theme attached property can not be meaningfully used form a
stateless JavaScript module without a context Item instance, because,
well, it is designed to be an attached property after all! It attaches
to an Item, and reads values based on a context and parent chain of it.

Instead, we have few options:

a) Turn the module into a stateful script. This is horrible for scripts
which are to be reused from more than one QML component, because they
rely on external imports from those QML files.

b) Pass the context Item to the function and rewrite the assignments to
query the attached properties from that Item. Sounds good and type-safe.

c) Return only names of the respective color properties, so that the
client code may query them on any number of any Items they like. This
is the least intrusive option, and does not fetch colors unnecessarily
for the DeviceView component which doesn't currently use it anyway.

BUG: 481216